Fix root=HEX in cmdline
A root= boot parameter can be specified by the number of the device or
by major and minor number. In this case we can't write it in /etc/fstab
without conversion because the mount(1) doesn't understand the follow
lines:

8:17 /newroot ext3 defaults 1 1
811  /newroot ext3 defaults 1 1

That why we have to convert it to mount-readable form. We can make the
conversion even before there will be real device files:

/dev/block/8:17 /newroot ext3 defaults 1 1

I want to say thanks to Sergey Vlasov for this good idea.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>