Refactor features rules
Most features copy files and directories from the system to the
initramfs. Because of this, there are a lot of put-file and put-tree
calls, which reduces performance. There is also the problem of
specifying the correct dependencies between targets.

Instead, we can accumulate a list of files and directories and copy them
to the initramfs in one place.

For these purposes, two variables PUT_FEATURE_DIRS and
PUT_FEATURE_FILES were created.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>