Refactor feature dependencies
Each feature can require and disable other features. There are several
issues with the implementation of this functionality.

1. Requires are implemented through recursive includes which is pretty
   slow.
2. Disabling other features occurs at the time of include. This means
   that if a disabled feature has already been included, then it can be
   disabled.

To solve this, you need to move the definition of dependencies to
config.mk which is always read and generate lists with dependencies
that will be used to calculate the list of used features.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>