Feature luks: decrypt using plymouth if present
When plymouth is present and running, systemd tries to decrypt LUKS
partition by invoking cryptsetup directly.

It looks like input goes to nowhere and cryptsetup is trying to decrypt
with empty password given.

Fedora uses 'plymouth ask-for-password' in that cases, making possible
to pass input to cryptsetup, decrypt volumes successfully and show a
prompt at splash screen, giving information that there's a passphrase
required to proceed with.