tests: Add test for incorrect configuration (root partition + luks partiton)
The test checks the sequence of uevent processing. In the test system,
the root partition is on a regular partition, but there is also
an encrypted partition. Despite the fact that the root partition does
not use luks, this feature has been added to the image.

Even in this configuration, the normal partition should be processed
first and there should be no password prompt for the luks partition.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>