runtime: ueventd tries to process events again if it did not work the first time
Sometime events may be received in a wrong order. This means that it is
impossible to process them in one pass through the cycle. If no new
event arrives after the last one, the processing may get stuck.

As a solution, the event handler sends an event to retry processing the
remaining events. The new iteration will accept for processing new
events that may appear between iterations.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>