Feature kbd: Require gpu-drm if framebuffer is used on vtconsole
This should fix the following issue: the console is in text mode, kbd
feature initializes the console font and the boot continues. The real
system loads framebuffer module and this breaks the console font setup.

A technically real system must reconfigure the console font after
loading framebuffer module, but not all systems seem to be able to do
this.

So if at the time of initramfs creation the framebuffer is used in
the console, then we add the feature gpu-drm feature to the image.

In this case, modules for the framebuffer will get into the initramfs,
they will be loaded from the initramfs and the console font will be
configured correctly.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>