Allow to compile and install feature-specific utilities
Some features require binary utilities. Before this the only solution to
this is to put them in datasrc/ and put them unconditionally.

Now it is possible to create a src/ directory in the features directory:

features/<name>/src/Makefile.mk

Inside Makefile.mk, utilities are described in the same way as in
datasrc/.In addition to the already existing placement paths,
$(dest_feature_bindir), $(dest_feature_sbindir) has been added. They
points to the directories in which $(PUT_FEATURE_PROGS) will also be
searched if the feature _enabled_ in configuration.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>