Feature gpu-drm: Filter enabled and/or connected drm devices
Use only enabled drm devices. But this is not enough because some NVIDIA
driver versions may not show which card is enabled. In this case, we
additionally check the status. So, we are looking for modules for cards
that are either enabled or a monitor is connected to them.

From https://forums.developer.nvidia.com/t/sys-class-drm-enabled-reports-disabled-for-enabled-monitors/51187
Subject: /sys/class/drm/*/enabled reports “disabled” for enabled monitors

Dear support,

I am writing this from a laptop with a GTX1070 on driver 375.66. The
internal panel is on eDP-1, while I have an external monitor connected
on DP-1.

$ grep . /sys/class/drm/*/status
/sys/class/drm/card0-DP-1/status:connected
/sys/class/drm/card0-DP-2/status:disconnected
/sys/class/drm/card0-eDP-1/status:connected
/sys/class/drm/card0-HDMI-A-1/status:disconnected

Yet these connected monitors are reported as disabled by the nvidia
driver:

$ grep . /sys/class/drm/*/enabled
/sys/class/drm/card0-DP-1/enabled:disabled
/sys/class/drm/card0-DP-2/enabled:disabled
/sys/class/drm/card0-eDP-1/enabled:disabled
/sys/class/drm/card0-HDMI-A-1/enabled:disabled

In fact these files always report "disabled", no matter what monitors I
connect/disconnect and what resolution I drive them at. This is not the
case for nouveau (or intel on another laptop), where "enabled" or
"disabled" are reported, reflecting the current monitor status.

The nasty consequence of this is that systemd suspends the laptop when I
close the lid, because it concludes that no external monitors are
connected.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>