Use separate sessions
The /dev/console is just one, but we potentially have several
users who can use the console in parallel. These are, for example, the
progress of loading services, initrd shell, utilities asking for a
password.

For local boot, we can use different vt to solve this problem, but that
won't work for serial and netconsole.

As a solution, we are using terminal multiplexer for the /dev/console.
The init creates the first bootlog session and switches between other
sessions if they appear.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>