Feature ucode: Fix firmware definition for amd cpu
For cpu vendor: AuthenticAMD, cpu family: 15 add-ucode does not work
properly because the code is always trying to get the hex value of the
cpu family and hex(15) == f.

We should not try to compare the resulting hex with a number. We just
need to check for the existence of the file in case there is a microcode
for this family.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>