Detect separate /usr partition
On modern systems, almost the entire installed operating system is
located in /usr and /bin, /sbin and /lib point inside it.

```
/
|-- etc
|-- usr
|   |-- bin
|   |-- sbin
|   |-- lib
|   `-- lib64
|-- run
|-- var
|-- bin -> usr/bin
|-- sbin -> usr/sbin
|-- lib -> usr/lib
`-- lib64 -> usr/lib64
```

If /usr is on a separate partition, we need to mount it in the same way
as / (root partition). We are ready for it. The only thing missing is
the auto-detection of this situation.

Link: https://fedoraproject.org/wiki/Features/UsrMove
Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>