Add udev rules parser
udevd does not have a public library. There is only the libudev library,
which is designed to work with a running daemon.

We need a parser of rules and the ability to analyze them. This parser
allows us to search for errors in the syntax of udev rules and also get
information about external utilities that are called from the rules.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>