feature/kickstart: Reset rootdelay timer after kickstart
If the rootdelay timer has started counting down, then after poll is
unpaused, a timeout may occur. This is because the rootdelay remembers
the time and if the kickstart took a long time it will timeout the next
time the rootdelay is checked.

To avoid this we reset the timer before we unpause poll.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>