feature/procacct: Add bpf helper
This bpf helper translates to userspace information about the arguments
to the execve system call. This is necessary because taskstats has no
other information about the process other than its comm. In the case of
shell scripts and some other cases, comm is not very informative.

Of course, in this approach, a race situation is possible between an
event from a bpf program and a netlink event. In this case, a memory
leak will occur because no one will remove the entry from the tree.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>