Use timestamp instead of date
When the rtc_rk808 kernel module is loaded, the time changes
dramatically from 1970 to the current one. This results in the rootdelay
being triggered.

To avoid this we need to use CLOCK_BOOTTIME (a nonsettable system-wide
clock that is identical to CLOCK_MONOTONIC, except that it also includes
any time that the system is suspended).

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>