feature/plymouth: Show splash in synchronous mode
Before this patch, splash display was done asynchronously. The
background process waited for the framebuffer to appear and only then
turned on splash. I can't remember who suggested this and why. Perhaps
the motivation was to not slow down the boot process. Maybe.

But this approach leads to a race condition. While the background
process is waiting for the framebuffer, luks may appear and ask the
password for partition. Since there is no splash yet, the password is
requested in the text console. And after that, the background process
turns on splash, overriding the password prompt. On a slow processor
this can be reproduced quite well.

I find it strange that after the service is started the plymouth is
not ready to go because splash is not shown. This breaks service startup
dependencies.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>