feature/pipeline: Allow to specify mounting parameters for opevelayfs
The overlayfs has a number of parameters (redirect_dir, index, etc.)
that are sometimes useful to use. To specify parameters, we extend the
format of the overlayfs= parameter:

overlayfs=[mountopt,mountopt,...:]pipe1[,pipe2,...]

Mount options and lowerdirs are separated by colon (":").

overlayfs=redirect_dir=on,index=on:pipe1,pipe2,pipe3

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>