feature/pipeline: Add support the "data-only" lower layers in overlayfs
Below the top most lower layer, any number of lower most layers may be
defined as "data-only" lower layers, using double comma (",,")
separators:

overlayfs=pipe1,,pipe2

The paths of files in the "data-only" lower layers are not visible in
the merged overlayfs directories and the metadata and st_ino/st_dev of
files in the "data-only" lower layers are not visible in overlayfs
inodes.

Only the data of the files in the "data-only" lower layers may be
visible when a "metacopy" file in one of the lower layers above it, has
a "redirect" to the absolute path of the "lower data" file in the
"data-only" lower layer.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>