plymouth: fix plymouth showing when drm is not ready yet
Fix waiting for a real video card when simpledrm is available.
card0 will disappear when the DRM module of the real video card is
loaded. Because of this, by the time Plymouth was shown, simpledrm
was no longer there, and there was no real video card yet.

Instead of waiting for devices to appear, is proposed to launch
triggers for subsystems, as is uses in dracut.

Signed-off-by: Anton Midyukov <antohami@altlinux.org>