Allow to set specific array elements
If a parameter is defined as an array (register_array), then multiple
occurrences of that parameter are combined into an array by appending.
Specifying parameters in /proc/cmdline:

  mountfs=aaa mountfs=bbb mountfs=ccc

will lead to the creation:

  MOUNTFS0=aaa
  MOUNTFS1=bbb
  MOUNTFS2=ccc
  MOUNTFS=3

This is only useful when all elements have values. If you need to
specify the value only for the last element, then you have to make ugly
constructions:

  mountfs= mountfs= mountfs=ccc

To avoid this, we can add the ability to specify a specific array
element:

  mountfs[2]=ccc

will lead to the creation:

  MOUNTFS0=
  MOUNTFS1=
  MOUNTFS2=ccc
  MOUNTFS=3

As a side effect, it becomes possible to specify elements in different
orders:

  mountfs[2]=ccc mountfs[0]=aaa

will create:

  MOUNTFS0=aaa
  MOUNTFS1=
  MOUNTFS2=ccc
  MOUNTFS=3

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>