Add special wrappers for findmnt and blkid
The system utilities findmnt and blkid work increasingly worse in an
isolated environment. In addition to accessing /proc, they try to check
for block devices even if the information is in their cache.

Since we don't have real devices in bugreport or testcases, this makes
it impossible to use findmnt in an emulated environment.

To solve this we will have to emulate the behavior of findmnt and blkid.
This is a terrible decision, but there are no other options yet.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>