kickstart: Rewrite block devices detection
Instead of having a whitelist of block device types you can try to guess
if it is possible to scan partitions. The same code is used in systemd.

Another improvement is that we switched to major:minor when excluding
devices. This will allow to ignore device names.

Also, block devices are sorted by PCI domain and BUS number instead of
device names.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>