make-initrd: Reorder arguments and make variables in the command line
Some users use the make feature that allows to override makefile
variables via command line.

In this case, the order in which arguments and variables are specified
is important because the make-initrd utility analyzes its arguments.

Correct order:

  make-initrd uki UKI_FEATURES=foo UKI_IMAGEFILE=/tmp/uki.efi

Bad order:

  make-initrd UKI_FEATURES=foo UKI_IMAGEFILE=/tmp/uki.efi uki

To avoid such issues it is necessary to reorder the arguments and move
the variables after the normal arguments.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>