guess/touchscreen: Fix the interpretation of the value in properties
The value in the /sys/class/input/event*/device/properties is in hex
format, not integer. So if the device has the following properties:

INPUT_PROP_POINTER | INPUT_PROP_BUTTONPAD | INPUT_PROP_SEMI_MT

will have a “d” in the */device/properties file. Such a value causes a
syntax error.

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/Documentation/input/event-codes.rst
Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>