Made a wrapper for ldconfig
Direct use of ldconfig is not portable. In musl, unlike glibc, there is
no ldconfig utility at all. In gentoo, debian and other distributions
that have a port to musl have their own utility that simulates ldconfig.

Unfortunately all such utilities are fully or partially incompatible
with ldconfig from glibc.

Therefore, instead of calling ldconfig directly, a wrapper will be used
that implements the necessary functionality. With glibc, it is possible
to use ldconfig directly. And for musl, emulation will be used.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>