feature/runtime: Use own utility to switch root
Since it is now possible to boot via compressed image and overlayfs,
support for the ability to unmount root in initramfs is now required.

Th switch_root from busybox or util-linux does not allow this. Moreover,
if initramfs is not on ramfs/tmpfs these utilities will not remove the
content and will leave it untouched.

So we need our own utility that allows us to unmount old rootfs if
possible.

As an added benefit, it will be possible to do away with unnecessary
actions when switching rootfs.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>