Add information about missing firmware
If the module in the initramfs depends on firmware that is not in
initramfs, it can potentially cause the device to malfunction or make it
impossible to boot the computer.

But since this is not necessarily a bug since the user may want to
remove “extra” firmware that is not used by the current hardware.

So just limit ourselves to the warning that it was not possible to
find all firmware for all kernel modules.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>