Include features/*/rules.mk only once
Second include of rules.mk causes shell and other functions to be
executed again. This adds complexity for the user when writing rules.mk.
He must know and remember that the written code will be executed more
than once. This is an unreasonable restriction.

To avoid this it is better to add a new file to affect the image
packaging process - pack.mk. The code in this file will also be called
only once.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>