Improve quoting performance
If features or users have added a large number of files or directories
to pack into an image, then quoting a very long string takes too much
time.

Quoting variables is necessary for the correct processing of paths with
spaces.

The problem is that if we need to process a very large string,
character-by-character parsing takes a long time. It so happens that
there is no quick way to parse a string character by character in the
POSIX shell. But in bash, such a way exists.

Since we always use bash, the quoting process can be significantly
accelerated by simply reading a single character in a string.

Before:

[   2.55ms ] Config file: /etc/initrd.mk
[     5.3m ] Used features: runtime add-modules add-udev-rules cleanup compress depmod-image gpu-drm locales plymouth rdshell rootfs system-glibc ucode
[     5.3m ] Unpacked size: 36M
[     5.3m ] Image size: 16M
[     5.3m ] Image is saved as /boot/initrd-6.16.0.img

After:

[   2.47ms ] Config file: /etc/initrd.mk
[    9.43s ] Used features: runtime add-modules add-udev-rules cleanup compress depmod-image gpu-drm locales plymouth rdshell rootfs system-glibc ucode
[    9.43s ] Unpacked size: 36M
[    9.44s ] Image size: 16M
[    9.49s ] Image is saved as /boot/initrd-6.16.0.img

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>