feature/mdadm: Guess modules for raid components
If the RAID is built on NVMe, then modules for the NVMe subsystem must
also be checked and added. The sysfs contains information about these
components.

As component devices are added to an md array, they appear in the `md`
directory as new directories named `dev-XXX` where `XXX` is a name that
the kernel knows for the device, e.g. hdb1.

Link: https://web.git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/Documentation/admin-guide/md.rst:470
Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>