feature/guess-nvme: New feature to add nvme modules
The feature adds the ability to search for and add nvme devices.
The nvme sysfs looks like this, where there are no modalias because
these are virtual nodes:

    /sys/devices/virtual/nvme-subsystem/nvme-subsys0/...

with the specific file `subsystype` with value `nvm`. In this case,
symlinks `nvme[0-9]` can be searched for, which point to the actual
block device.

Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>